# openapi.yaml
openapi: 3.1.0
info:
  title: Predictify API
  version: 0.0.1
  description: Backend API for Predictify — a Stellar/Soroban prediction-markets dApp
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            requestId:
              type: string
          required:
            - code
      required:
        - error
    ValidationErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            details:
              nullable: true
          required:
            - code
      required:
        - error
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        stellarAddress:
          type: string
        score:
          type: number
      required:
        - rank
        - stellarAddress
        - score
...
    RateLimitAuditEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
          enum:
            - rate_limit.blocked
        walletAddress:
          type: string
          nullable: true
        ip:
          type: string
        correlationId:
          type: string
        rateLimitContext:
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - walletAddress
        - ip
        - correlationId
        - createdAt
...
    FeatureFlag:
      type: object
      properties:
        key:
          type: string
        enabled:
          type: boolean
        description:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - key
        - enabled
        - description
        - updatedAt
    FeatureFlagListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FeatureFlag'
      required:
        - data
    FeatureFlagResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/FeatureFlag'
      required:
        - data
    CreateFeatureFlagRequest:
      type: object
      properties:
        key:
          type: string
        enabled:
          type: boolean
        description:
          type: string
          nullable: true
          maxLength: 280
      required:
        - key
        - enabled
...
    UserListRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stellarAddress:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - stellarAddress
        - createdAt
...
    PredictionRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        marketId:
          type: string
        question:
          type: string
        outcome:
          type: string
        amount:
          type: string
        txHash:
          type: string
        status:
          type: string
          enum: *ref_0
        result:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        resolutionTime:
          type: string
          format: date-time
      required:
        - id
        - marketId
        - question
        - outcome
        - amount
        - txHash
        - status
        - result
        - createdAt
        - resolutionTime
    PredictionsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PredictionRow'
        nextCursor:
          type: string
          nullable: true
      required:
        - data
        - nextCursor
...
    AuditEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        actor:
          type: string
        targetAddress:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - createdAt
...
    AuditCountsSummary:
      type: object
      properties:
        totalCount:
          type: integer
        byAction:
          type: array
          items:
            $ref: '#/components/schemas/AuditActionCount'
      required:
        - totalCount
        - byAction
...
    WebhookDelivery:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique delivery ID
        eventId:
          type: string
          description: Opaque event identifier supplied by the emitting service
        eventType:
          type: string
          description: Event type string, e.g. "market.resolved" or "dispute.opened"
        targetUrl:
          type: string
          format: uri
          description: The subscriber endpoint the delivery is sent to
        payloadBase64:
          type: string
          description: Base64-encoded signed request body
        signature:
          type: string
          description: HMAC-SHA256 signature over the payload, sent as a request header
        headers:
          type: object
          nullable: true
          additionalProperties:
            type: string
          description: Extra HTTP headers sent with the delivery (may be null)
        status:
          $ref: '#/components/schemas/DeliveryStatus'
        attempts:
          type: integer
          minimum: 0
          description: Number of delivery attempts made so far
        maxAttempts:
          type: integer
          minimum: 0
          exclusiveMinimum: true
          description: Maximum attempts before the delivery is dead-lettered
        lastError:
          type: string
          nullable: true
          description: Error message from the most recent failed attempt, or null
        nextAttemptAt:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 timestamp of the next scheduled retry, or null when terminal
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when this delivery record was created
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the most recent status change
      required:
        - id
        - eventId
        - eventType
        - targetUrl
        - payloadBase64
        - signature
        - headers
        - status
        - attempts
        - maxAttempts
        - lastError
        - nextAttemptAt
        - createdAt
        - updatedAt
...
paths:
...
  /api/leaderboard:
    get:
      operationId: getLeaderboard
      tags:
        - Leaderboard
      summary: Get global leaderboard
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            maximum: 100
            default: 50
          required: false
          name: limit
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
            default: 0
          required: false
          name: offset
          in: query
        - schema:
            type: boolean
            nullable: true
            default: false
          required: false
          name: refresh
          in: query
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
                  meta:
                    type: object
                    properties:
                      limit:
                        type: number
                      offset:
                        type: number
                      count:
                        type: number
                      refresh:
                        type: boolean
                    required:
                      - limit
                      - offset
                      - count
                      - refresh
                required:
                  - data
                  - meta
...
  /api/rate-limit:
    get:
      operationId: listRateLimitEvents
      tags:
        - Rate Limiting
      summary: List rate-limit audit events (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated rate-limit audit log
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RateLimitAuditEntry'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/admin/feature-flags:
    get:
      operationId: listAdminFeatureFlags
      tags:
        - Admin
      summary: List configured feature flags
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of feature flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    post:
      operationId: createAdminFeatureFlag
      tags:
        - Admin
      summary: Create a new feature flag
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureFlagRequest'
      responses:
        '201':
          description: Feature flag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Feature flag already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/users:
    get:
      operationId: listUsers
      tags:
        - Users
      summary: List all users (cursor-paginated)
      description: >-
        Returns a cursor-paginated list of registered users sorted newest-first (DESC createdAt, DESC id).  Pass the
        opaque `nextCursor` value from one response as the `cursor` query parameter of the next request to advance
        through pages.  A null `nextCursor` indicates the last page.
      parameters:
        - schema:
            type: string
            description: Opaque cursor token from the previous page's nextCursor field.
            example: djF8MjZ8MjAyNi0wMS0wMVQwMDowMDowMC4wMDBafGFiY2Qtd...
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            description: Page size (1–100, default 20).
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserListRow'
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor for the next page, or null on the last page.
                required:
                  - data
                  - nextCursor
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
...
  /api/predictions:
    get:
      operationId: listPredictions
      tags:
        - Predictions
      summary: List the authenticated user’s predictions
      description: >-
        Returns a cursor-paginated list of predictions placed by the caller. Sort order is `createdAt DESC, id DESC`.
        Pass the returned `nextCursor` as `?cursor=` to fetch the next page. `nextCursor` is `null` when no further
        pages exist.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 128
          required: false
          name: marketId
          in: query
        - schema:
            type: string
            enum: *ref_0
          required: false
          name: status
          in: query
        - schema:
            type: string
            minLength: 1
            maxLength: 64
          required: false
          name: outcome
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated list of predictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionsListResponse'
              examples:
                authenticatedPredictionsPage:
                  value:
                    data:
                      - id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                        marketId: market_123
                        question: Will Bitcoin hit 100k in 2026?
                        outcome: 'yes'
                        amount: '50.0000000'
                        txHash: 8c253240be423ef8109d94101e40a02bc8f297b819f0ff4f4c20b8e906059e66
                        status: won
                        result: 'yes'
                        createdAt: '2026-05-01T12:00:00.000Z'
                        resolutionTime: '2026-06-01T12:00:00.000Z'
                    nextCursor: cursor_abc123
        '400':
          description: Validation error — invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
              examples:
                invalidLimit:
                  value:
                    error:
                      code: VALIDATION_ERROR
                      details: Limit must be between 1 and 100
        '401':
          description: Unauthorized — missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                unauthorized:
                  value:
                    error:
                      code: UNAUTHORIZED
                      requestId: req_xyz789
...
  /api/admin/audit:
    get:
      operationId: getAdminAuditLog
      tags:
        - Admin
      summary: List audit log entries (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: false
          name: action
          in: query
        - schema:
            type: string
          required: false
          name: actor
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: startDate
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: endDate
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/audit/counts:
    get:
      operationId: getAuditCounts
      tags:
        - Admin
      summary: Per-action audit log counts summary for dashboards (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: date-time
          required: false
          name: startDate
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: endDate
          in: query
      responses:
        '200':
          description: Audit log counts summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuditCountsSummary'
                required:
                  - data
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/webhooks:
    get:
      operationId: listWebhookDeliveries
      tags:
        - Webhooks
      summary: List live webhook deliveries (admin only)
      description: >-
        Returns a cursor-paginated list of live webhook delivery records ordered newest-first (by `createdAt` DESC, then
        `id` DESC as a stable tie-breaker). Only deliveries that have not yet been dead-lettered are returned here;
        exhausted deliveries appear in the DLQ at `GET /api/admin/webhooks/dlq`. Requires an admin JWT (`role:
        "admin"`).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            minLength: 1
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            maximum: 100
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated page of live webhook deliveries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookDelivery'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
              examples:
                webhookDeliveriesPage:
                  summary: First page of live deliveries with one pending and one delivered record
                  value:
                    data:
                      - id: d1a2b3c4-0001-4000-8000-000000000001
                        eventId: evt-market-resolved-001
                        eventType: market.resolved
                        targetUrl: https://example.com/hooks/predictify
                        payloadBase64: >-
                          eyJldmVudCI6Im1hcmtldC5yZXNvbHZlZCIsImlkIjoiZDFhMmIzYzQtMDAwMS00MDAwLTgwMDAtMDAwMDAwMDAwMDAxIn0=
                        signature: sha256=abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890
                        headers: null
                        status: delivered
                        attempts: 1
                        maxAttempts: 5
                        lastError: null
                        nextAttemptAt: null
                        createdAt: '2026-07-25T12:00:00.000Z'
                        updatedAt: '2026-07-25T12:00:05.000Z'
                      - id: d1a2b3c4-0002-4000-8000-000000000002
                        eventId: evt-dispute-opened-001
                        eventType: dispute.opened
                        targetUrl: https://example.com/hooks/predictify
                        payloadBase64: eyJldmVudCI6ImRpc3B1dGUub3BlbmVkIn0=
                        signature: sha256=fedcba0987654321fedcba0987654321fedcba0987654321fedcba0987654321
                        headers: null
                        status: pending
                        attempts: 0
                        maxAttempts: 5
                        lastError: null
                        nextAttemptAt: '2026-07-25T11:00:30.000Z'
                        createdAt: '2026-07-25T11:00:00.000Z'
                        updatedAt: '2026-07-25T11:00:00.000Z'
                    nextCursor: >-
                      eyJzb3J0VmFsdWUiOiIyMDI2LTA3LTI1VDExOjAwOjAwLjAwMFoiLCJpZCI6ImQxYTJiM2M0LTAwMDItNDAwMC04MDAwLTAwMDAwMDAwMDAwMiJ9
                emptyPage:
                  summary: Empty page when no live deliveries exist
                  value:
                    data: []
                    nextCursor: null
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                invalidLimit:
                  summary: Non-numeric limit value
                  value:
                    error:
                      code: validation_error
                      message: limit must be a positive integer
                      requestId: req-abc-123
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Caller is not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...